Removing panda redirect
