--- app/com/gu/viewer/views/login.scala.html
@@ -1,5 +1,0 @@
-@(domain: String)
-
-<div class="container">
-    <h1>Please sign in to <a href="https://login.@domain/login?returnUrl=https://viewer.@domain/">login.gutools</a> to access viewer</h1>
-</div>
