+++ conf/routes
@@ -16,8 +16,7 @@ GET        /live/*path                    com.gu.viewer.controllers.Application.
 GET        /proxy/preview/*path           com.gu.viewer.controllers.Proxy.proxy(service="preview", path)
 GET        /proxy/live/*path              com.gu.viewer.controllers.Proxy.proxy(service="live", path)
 
-GET        /oauthCallback                 com.gu.viewer.controllers.Login.oauthCallback
-POST       /send-email                    com.gu.viewer.controllers.Application.sendEmail(path)
+POST       /send-email                    com.gu.viewer.controllers.Email.sendEmail(path)
 
 
 # Long life cache for statically versioned font assets
